+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 =============
 * booking.com: pull old reservations (redmine #590)
 * booking.com: errors && logs (redmine #595)
+* booking.com: auto tariffs fix (redmine #608)
 
 1.5.5
 =============