+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * booking.com: pull old reservations (redmine #590)
 * booking.com: errors && logs (redmine #595)
 * booking.com: auto tariffs fix (redmine #608)
+* booking.com: config validation (redmine #642)
 
 1.5.5
 =============